pyproject.toml
[tool.poetry]
name = "instaclone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
pyjwt = "^2.10.1"
pydantic-settings = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
